reworking to split interface and impl - not yet compiling
