+++ java/modules/math/src/main/java/module-info.java
@@ -11,8 +11,8 @@
  * WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-module moreland.sample.modules.math {
+module mathservice {
     requires java.base;
-    //requires moreland.sample.modules.mathservice;
+
     exports moreland.sample.modules.math;
 }
