+++ java/modules/pom.xml
@@ -4,13 +4,13 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>moreland.samples.modules</groupId>
+    <groupId>moreland.sample.modules</groupId>
     <artifactId>modules</artifactId>
     <packaging>pom</packaging>
     <version>1.0-SNAPSHOT</version>
     <modules>
-        <module>mathservice</module>
         <module>math</module>
+        <module>math-simple</module>
     </modules>
 
     <properties>
